Fixed small bug. see tracker 2119
git-svn-id: https://opencmiss.svn.sourceforge.net/svnroot/opencmiss/cm/trunk@864 61e36070-c886-4827-b7c0-ed9e4dc59cc8

+++ src/opencmiss_c.f90
@@ -11521,7 +11521,7 @@ END FUNCTION CMISSFieldsTypeCreateC
     INTEGER(C_INT) :: CMISSFieldParameterSetDataRestoreLCPtr !<Error Code.
     !Local variables
     TYPE(CMISSFieldType), POINTER :: Field
-    REAL(C_INT), POINTER :: Parameters(:)
+    INTEGER(C_INT), POINTER :: Parameters(:)
 
     CMISSFieldParameterSetDataRestoreLCPtr = CMISSNoError
     IF(C_ASSOCIATED(ParametersPtr)) THEN
